Remove TileWidth and TileLength tags
When these are included, we should really be including the TileOffsets
and TileByteCounts tags as well, otherwise the resulting behaviour can
be undefined. But in our case, we are not storing images in tiled
format, so it is easier just to leave them out.

This fixes DNG reading for RawTherapee 5.11.